+++ src/main/java/com/fiap/fast_food_tc/domain/usecase/impl/CustomerUseCaseImpl.java
@@ -1,6 +1,6 @@
 package com.fiap.fast_food_tc.domain.usecase.impl;
 
-import com.fiap.fast_food_tc.cross.CustomerMapper;
+import com.fiap.fast_food_tc.cross.mapper.CustomerMapper;
 import com.fiap.fast_food_tc.domain.entity.ECustomer;
 import com.fiap.fast_food_tc.domain.gateway.CustomerGateway;
 import com.fiap.fast_food_tc.domain.usecase.CustomerUseCase;
@@ -34,16 +34,16 @@ public class CustomerUseCaseImpl implements CustomerUseCase {
 
     @Override
     public void update(ECustomer user) {
-
+        //TODO
     }
 
     @Override
-    public void delete(Long id) {
-
+    public void delete(Integer id) {
+        //TODO
     }
 
     @Override
-    public ECustomer getById(Long id) {
+    public ECustomer getById(Integer id) {
         return null;
     }
